+++ lib/phoenix_replay/assets.ex
@@ -9,6 +9,16 @@ defmodule PhoenixReplay.Assets do
     conn
     |> put_resp_content_type("application/javascript")
     |> put_resp_header("cache-control", "public, max-age=3600")
-    |> send_resp(200, PhoenixReplay.Layouts.player_js())
+    |> send_resp(200, player_js())
+  end
+
+  defp player_js do
+    :code.priv_dir(:phoenix_replay)
+    |> Path.join("static/phoenix_replay/player.js")
+    |> File.read!()
+  rescue
+    _ in [ArgumentError, ErlangError, File.Error] ->
+      Path.join([File.cwd!(), "priv", "static", "phoenix_replay", "player.js"])
+      |> File.read!()
   end
 end